qcommon: frame pacing, noreturn errors and bounds fixes
- sleep between frames instead of busy waiting (com_busyWait keeps the
  old behaviour), separate com_maxfpsUnfocused/com_maxfpsMinimized
- Sys_Milliseconds uses CLOCK_MONOTONIC so wall clock changes don't
  warp the game time
- mark Com_Error (and friends) noreturn via Q_NORETURN
- Huff_offsetReceive stops at the end of the message instead of
  reading past it; MSG_ReadBits handles negative and 32 bit widths
- CM_TraceThroughSphere used a 'long' to alias a float, reading past
  it on LP64
- CM_SubdivideGridColumns stops when the grid is full
- FS_CopyFile checks allocations and frees its buffer on all paths
- vsprintf -> vsnprintf in the printf-style helpers

+++ src/qcommon/cm_trace.c
@@ -143,13 +143,13 @@ SquareRootFloat
 ================
 */
 float SquareRootFloat( float number ) {
-	long i;
+	int i;      // must be a 32-bit type: 'long' reads past the float on LP64
 	float x, y;
 	const float f = 1.5F;
 
 	x = number * 0.5F;
 	y  = number;
-	i  = *( long * ) &y;
+	i  = *( int * ) &y;
 	i  = 0x5f3759df - ( i >> 1 );
 	y  = *( float * ) &i;
 	y  = y * ( f - ( x * y * y ) );
